+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2025-12-16
+
+### Added
+
+- Add `nearly_eq` and `get_message` method for `TestPanicResult`.
+- Add `Eq` and `PartialEq` trait implementation for `TestPanicResult`.
+- Minor refactoring.
+
 ## [0.4.1] - 2025-07-14
 
 ### Changed
@@ -30,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Minor refactoring.
 
+[0.5.0]: https://github.com/nossie531/test_panic/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/nossie531/test_panic/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/nossie531/test_panic/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/nossie531/test_panic/compare/v0.3.0...v0.3.1